requirements.txt
requests==2.31.0
fake-headers==1.0.2
loguru==0.7.2
PyYAML==6.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
aiofiles==23.2.1
tqdm==4.66.1